T: test cleanup
